Chnaged building_color to floats in typeDefs.js

+++ server/schemas/typeDefs.js
@@ -21,9 +21,9 @@ const typeDefs = gql`
         building_position_x: Int
         building_position_y: Int
         building_position_z: Int
-        building_color_r: Int
-        building_color_g: Int
-        building_color_b: Int
+        building_color_r: Float
+        building_color_g: Float
+        building_color_b: Float
     }
 
     type Auth {
